LPS-77522 Rename journal-content-asset-addon-entry-common to journal-content-asset-addon-entry-api
