+++ modules/apps/web-experience/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Journal Content Asset Addon Entry API
+Bundle-SymbolicName: com.liferay.journal.content.asset.addon.entry.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.journal.content.asset.addon.entry
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Web Content
