+++ modules/apps/web-experience/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-api/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+}
