+++ modules/apps/web-experience/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-api/src/main/java/com/liferay/journal/content/asset/addon/entry/ContentMetadataAssetAddonEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.content.asset.addon.entry.common;
+package com.liferay.journal.content.asset.addon.entry;
 
 import com.liferay.portal.kernel.servlet.taglib.ui.AssetAddonEntry;
