+++ modules/apps/web-experience/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-api/src/main/resources/com/liferay/journal/content/asset/addon/entry/packageinfo
