--- modules/apps/web-experience/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-common/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Journal Content Asset Addon Entry Common
-Bundle-SymbolicName: com.liferay.journal.content.asset.addon.entry.common
-Bundle-Version: 2.0.4
-Export-Package: com.liferay.journal.content.asset.addon.entry.common
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Web Content
