--- modules/apps/web-experience/journal/journal-content-asset-addon-entry/journal-content-asset-addon-entry-common/build.gradle
@@ -1,5 +1,0 @@
-dependencies {
-	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
-	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	provided project(":apps:web-experience:journal:journal-service")
-}
